Bump all versions, fix deprecated ServiceContext and add ChromaDB disk persistence

+++ process_document.py
@@ -7,13 +7,14 @@ import chromadb
 import chromadb.errors as chromadb_errors
 from transformers import TRANSFORMERS_CACHE
 
-from embed_and_retrieve import get_logger
+from embed_and_retrieve import get_logger, CHROMA_DIR
 
 logger = get_logger()
 
 # Create a temporary directory to store uploaded files
 UPLOAD_DIR = "data"
 os.makedirs(UPLOAD_DIR, exist_ok=True)
+os.makedirs(CHROMA_DIR, exist_ok=True)
 
 def upload_file(uploaded_file):
     with tempfile.NamedTemporaryFile(delete=False, suffix=f".{uploaded_file.name.split('.')[-1]}", dir=UPLOAD_DIR) as tmp_file:
@@ -22,7 +23,11 @@ def upload_file(uploaded_file):
 
 def cleanup():
     # Remove temporary files
-    shutil.rmtree(UPLOAD_DIR, ignore_errors=True)
+    if os.path.exists(UPLOAD_DIR):
+        shutil.rmtree(UPLOAD_DIR, ignore_errors=True)
+
+    if os.path.exists(CHROMA_DIR):
+        shutil.rmtree(CHROMA_DIR, ignore_errors=True)
 
     # Clear Transformers cache (Remove any HF models downloaded)
     if os.path.exists(TRANSFORMERS_CACHE):
